requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
annotated-types=0.6.0=py39hecd8cb5_0
anyio=4.2.0=py39hecd8cb5_0
blas=1.0=mkl
bottleneck=1.3.7=py39h7b7cdfe_0
brotli=1.0.9=h6c40b1e_8
brotli-bin=1.0.9=h6c40b1e_8
ca-certificates=2024.3.11=hecd8cb5_0
certifi=2024.6.2=py39hecd8cb5_0
click=8.1.7=py39hecd8cb5_0
contourpy=1.2.0=py39ha357a0b_0
cycler=0.11.0=pyhd3eb1b0_0
distro=1.9.0=py39hecd8cb5_0
exceptiongroup=1.2.0=py39hecd8cb5_0
fonttools=4.51.0=py39h6c40b1e_0
freetype=2.12.1=hd8bbffd_0
h11=0.14.0=py39hecd8cb5_0
httpcore=1.0.2=py39hecd8cb5_0
httpx=0.26.0=py39hecd8cb5_0
idna=3.7=py39hecd8cb5_0
importlib_resources=6.1.1=py39hecd8cb5_1
intel-openmp=2023.1.0=ha357a0b_43548
jpeg=9e=h6c40b1e_1
kiwisolver=1.4.4=py39hcec6c5f_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=h6c40b1e_8
libbrotlidec=1.0.9=h6c40b1e_8
libbrotlienc=1.0.9=h6c40b1e_8
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_1
libffi=3.4.4=hecd8cb5_1
libpng=1.6.39=h6c40b1e_0
libtiff=4.5.1=hcec6c5f_0
libwebp-base=1.3.2=h6c40b1e_0
lz4-c=1.9.4=hcec6c5f_1
matplotlib=3.8.4=py39hecd8cb5_0
matplotlib-base=3.8.4=py39hb47e01b_0
mkl=2023.1.0=h8e150cf_43560
mkl-service=2.4.0=py39h6c40b1e_1
mkl_fft=1.3.8=py39h6c40b1e_0
mkl_random=1.2.4=py39ha357a0b_0
ncurses=6.4=hcec6c5f_0
numexpr=2.8.7=py39h827a554_0
numpy=1.26.4=py39h827a554_0
numpy-base=1.26.4=py39ha186be2_0
openai=1.25.0=py39hecd8cb5_0
openjpeg=2.4.0=h66ea3da_0
openssl=3.0.13=hca72f7f_2
packaging=23.2=py39hecd8cb5_0
pandas=2.2.1=py39hcec6c5f_0
pillow=10.3.0=py39h6c40b1e_0
pip=24.0=py39hecd8cb5_0
pydantic=2.5.3=py39hecd8cb5_0
pydantic-core=2.14.6=py39hf2ad997_0
pyparsing=3.0.9=py39hecd8cb5_0
python=3.9.18=h5ee71fb_0
python-dateutil=2.9.0post0=py39hecd8cb5_2
python-dotenv=0.21.0=py39hecd8cb5_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py39hecd8cb5_0
readline=8.2=hca72f7f_0
seaborn=0.12.2=py39hecd8cb5_0
setuptools=69.5.1=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py39hecd8cb5_0
sqlite=3.45.3=h6c40b1e_0
tbb=2021.8.0=ha357a0b_0
tk=8.6.14=h4d00af3_0
tornado=6.3.3=py39h6c40b1e_0
tqdm=4.66.4=py39h20db666_0
typing-extensions=4.11.0=py39hecd8cb5_0
typing_extensions=4.11.0=py39hecd8cb5_0
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py39h6c40b1e_0
wheel=0.43.0=py39hecd8cb5_0
xz=5.4.6=h6c40b1e_1
zipp=3.17.0=py39hecd8cb5_0
zlib=1.2.13=h4b97444_1
zstd=1.5.5=hc035e20_2
